Sun, Oct 22, 2023, 10:52 PM -03:00

+++ credit/credit.c
@@ -17,7 +17,7 @@ int main(void)
     printf(luhn(creditCard) ? "true \n" : "false \n");
     printf("Card has %i digits! \n", countDigits(creditCard));
 
-    if (countDigits(creditCard) == 15 && (creditCard % pow(10,13) == 34 || creditCard % pow(10,13) == 37)){
+    if (countDigits(creditCard) == 15 && (creditCard % (long)pow(10,13) == 34 || creditCard % (long)pow(10,13) == 37)){
         printf("AMEX\n");
     }
 }
